pull out GranularLogHandler from SOTA into its own file
