+++ inbm/dispatcher-agent/tests/unit/sota/test_sota.py
@@ -9,7 +9,7 @@ from typing import Dict, List
 from dispatcher.packagemanager.memory_repo import MemoryRepo
 from dispatcher.sota.command_list import CommandList
 from dispatcher.sota.os_factory import SotaError, SotaOsFactory
-from dispatcher.sota.sota import SOTA
+from dispatcher.sota.sota import SOTA, GranularLogHandler
 from dispatcher.sota.sota import SOTAUtil
 from dispatcher.sota.constants import *
 from inbm_lib.xmlhandler import XmlHandler
@@ -272,105 +272,3 @@ Processing triggers for man-db (2.9.1-1) ...
 Processing triggers for systemd (245.4-4ubuntu3.23) ...
         """
         self.assertFalse(TestSota.sota_instance._is_ota_no_update_available(cmd_list))
-
-
-    @patch('json.dump')
-    @patch('json.load', return_value={"UpdateLog":[]})
-    @patch('dispatcher.sota.sota.get_os_version', return_value='2.0.20240802.0213')
-    @patch('inbm_common_lib.shell_runner.PseudoShellRunner.run', return_value=("tiber", "", 0))
-    def test_save_granular_in_tiberos_with_success_log(self, mock_run, mock_get_os_version, mock_load, mock_dump) -> None:
-        TestSota.sota_instance.factory = SotaOsFactory(
-            TestSota.mock_disp_broker,
-            None, []).get_os('tiber')
-
-        TestSota.sota_instance._update_logger.detail_status = OTA_SUCCESS
-
-        with patch('builtins.open', mock_open()) as m_open:
-            TestSota.sota_instance.save_granular_log(check_package=False)
-
-        expected_content = {
-            "UpdateLog": [
-                {
-                    "StatusDetail.Status": OTA_SUCCESS,
-                    "Version": '2.0.20240802.0213'
-                }
-            ]
-        }
-
-        mock_dump.assert_called_with(expected_content, m_open(), indent=4)
-
-
-    @patch('json.dump')
-    @patch('json.load', return_value={"UpdateLog":[]})
-    @patch('dispatcher.sota.sota.get_os_version', return_value='2.0.20240802.0213')
-    @patch('inbm_common_lib.shell_runner.PseudoShellRunner.run', return_value=("tiber", "", 0))
-    def test_save_granular_in_tiberos_with_pending_log(self, mock_run, mock_get_os_version, mock_load, mock_dump) -> None:
-        TestSota.sota_instance.factory = SotaOsFactory(
-            TestSota.mock_disp_broker,
-            None, []).get_os('tiber')
-
-        TestSota.sota_instance._update_logger.detail_status = OTA_PENDING
-
-        with patch('builtins.open', mock_open()) as m_open:
-            TestSota.sota_instance.save_granular_log(check_package=False)
-
-        expected_content = {
-            "UpdateLog": [
-                {
-                    "StatusDetail.Status": OTA_PENDING,
-                    "Version": '2.0.20240802.0213'
-                }
-            ]
-        }
-
-        mock_dump.assert_called_with(expected_content, m_open(), indent=4)
-
-    @patch('json.dump')
-    @patch('json.load', return_value={"UpdateLog":[]})
-    @patch('inbm_common_lib.shell_runner.PseudoShellRunner.run', return_value=("tiber", "", 0))
-    def test_save_granular_in_tiberos_with_fail_log(self, mock_run, mock_load, mock_dump) -> None:
-        TestSota.sota_instance.factory = SotaOsFactory(
-            TestSota.mock_disp_broker,
-            None, []).get_os('tiber')
-
-        TestSota.sota_instance._update_logger.detail_status = FAIL
-        TestSota.sota_instance._update_logger.error = 'Error getting artifact size from https://registry-rs.internal.ledgepark.intel.com/v2/one-intel-edge/tiberos/manifests/latest using token'
-
-        with patch('builtins.open', mock_open()) as m_open:
-            TestSota.sota_instance.save_granular_log(check_package=False)
-
-        expected_content = {
-            "UpdateLog": [
-                {
-                    "StatusDetail.Status": FAIL,
-                    "FailureReason": 'Error getting artifact size from https://registry-rs.internal.ledgepark.intel.com/v2/one-intel-edge/tiberos/manifests/latest using token'
-                }
-            ]
-        }
-
-        mock_dump.assert_called_with(expected_content, m_open(), indent=4)
-
-
-    @patch('json.dump')
-    @patch('json.load', return_value={"UpdateLog":[]})
-    @patch('inbm_common_lib.shell_runner.PseudoShellRunner.run', return_value=("tiber", "", 0))
-    def test_save_granular_in_tiberos_with_rollback_log(self, mock_run, mock_load, mock_dump) -> None:
-        TestSota.sota_instance.factory = SotaOsFactory(
-            TestSota.mock_disp_broker,
-            None, []).get_os('tiber')
-
-        TestSota.sota_instance._update_logger.detail_status = ROLLBACK
-        TestSota.sota_instance._update_logger.error = 'FAILED INSTALL: System has not been properly updated; reverting..'
-        with patch('builtins.open', mock_open()) as m_open:
-            TestSota.sota_instance.save_granular_log(check_package=False)
-
-        expected_content = {
-            "UpdateLog": [
-                {
-                    "StatusDetail.Status": ROLLBACK,
-                    "FailureReason": 'FAILED INSTALL: System has not been properly updated; reverting..'
-                }
-            ]
-        }
-
-        mock_dump.assert_called_with(expected_content, m_open(), indent=4)
